+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 6.10.0
 * Add `SEPADirectDebitAccount` payment method
 * Add `SEPADirectDebitAccountDetails` to transaction object
 * Add `SEPA_DIRECT_DEBIT_ACCOUNT` to payment instrument type